수정: 레디스 추상화 및 불필요한 주석 삭제
