+++ src/main/kotlin/com/teamsparta/moamoa/infra/redis/RedissonLockManagerImpl.kt
@@ -1,0 +1,39 @@
+package com.teamsparta.moamoa.infra.redis
+
+import org.redisson.api.RLock
+import org.redisson.api.RedissonClient
+import org.slf4j.LoggerFactory
+import org.springframework.stereotype.Component
+import java.util.concurrent.TimeUnit
+
+@Component
+class RedissonLockManagerImpl(private val redissonClient: RedissonClient): RedissonLockManager {
+    private val log = LoggerFactory.getLogger(RedissonLockManagerImpl::class.java)
+
+    override fun acquireLock(
+        lockKey: String,
+        waitTime: Long,
+        leaseTime: Long,
+    ): Boolean {
+        val lock: RLock = redissonClient.getLock(lockKey)
+
+        try {
+            val lockable = lock.tryLock(waitTime, leaseTime, TimeUnit.MILLISECONDS)
+            if (!lockable) {
+                log.info("Lock 획득 실패={}", lockKey)
+                return false
+            }
+            log.info("로직 수행")
+            return true
+        } catch (e: InterruptedException) {
+            log.info("에러 발생")
+            throw e
+        }
+    }
+
+    override fun releaseLock(lockKey: String) {
+        val lock: RLock = redissonClient.getLock(lockKey)
+        log.info("락 해제")
+        lock.unlock()
+    }
+}
